Use separate update DAO class for new schema indexing
 This is required because the same connection was used to write in two transactions concurrently and jdbi was not happy with it.

 created separate jdbi managed class so that the connection is different and can work concurrently.
