+++ src/main/java/uk/gov/indexer/Indexer.java
@@ -4,6 +4,7 @@ import org.skife.jdbi.v2.DBI;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.gov.indexer.dao.DestinationDBUpdateDAO;
+import uk.gov.indexer.dao.DestinationDBUpdateDAO_NewSchema;
 import uk.gov.indexer.dao.IndexedEntriesUpdateDAO;
 import uk.gov.indexer.dao.SourceDBQueryDAO;
 import uk.gov.indexer.monitoring.CloudwatchRecordsProcessedUpdater;
@@ -17,6 +18,7 @@ public class Indexer {
     private final Configuration configuration;
     private final String register;
     private DestinationDBUpdateDAO destinationDBUpdateDAO;
+    private DestinationDBUpdateDAO_NewSchema destinationDBUpdateDAO_NewSchema;
     private SourceDBQueryDAO sourceDBQueryDAO;
 
     public Indexer(Configuration configuration, String register) {
@@ -30,6 +32,7 @@ public class Indexer {
 
             DBI destDbi = new DBI(configuration.getProperty(register + ".destination.postgres.db.connectionString"));
             destinationDBUpdateDAO = destDbi.open().attach(DestinationDBUpdateDAO.class);
+            destinationDBUpdateDAO_NewSchema = destDbi.open().attach(DestinationDBUpdateDAO_NewSchema.class);
 
             DBI sourceDbi = new DBI(configuration.getProperty(register + ".source.postgres.db.connectionString"));
             sourceDBQueryDAO = sourceDbi.onDemand(SourceDBQueryDAO.class);
@@ -51,7 +54,7 @@ public class Indexer {
                     new IndexEntryItemTask(
                             register,
                             sourceDBQueryDAO,
-                            destinationDBUpdateDAO
+                            destinationDBUpdateDAO_NewSchema
                     ),
                     0,
                     10,
@@ -86,6 +89,9 @@ public class Indexer {
     public synchronized void stop() {
         if (destinationDBUpdateDAO != null) {
             destinationDBUpdateDAO.close();
+        }
+        if (destinationDBUpdateDAO_NewSchema != null) {
+            destinationDBUpdateDAO_NewSchema.close();
         }
         if (sourceDBQueryDAO != null) {
             sourceDBQueryDAO.close();
