+++ src/main/java/uk/gov/indexer/dao/DestinationDBUpdateDAO.java
@@ -11,15 +11,12 @@ import org.skife.jdbi.v2.sqlobject.mixins.GetHandle;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.stream.Collectors;
 
 
 public abstract class DestinationDBUpdateDAO implements GetHandle, DBConnectionDAO {
     private final CurrentKeysUpdateDAO currentKeysUpdateDAO;
     private final IndexedEntriesUpdateDAO indexedEntriesUpdateDAO;
-    private final EntryUpdateDAO entryUpdateDAO;
-    private final ItemUpdateDAO itemUpdateDAO;
 
     public DestinationDBUpdateDAO() {
         Handle handle = getHandle();
@@ -33,21 +30,11 @@ public abstract class DestinationDBUpdateDAO implements GetHandle, DBConnectionD
         if (!indexedEntriesUpdateDAO.indexedEntriesIndexExists()) {
             indexedEntriesUpdateDAO.createIndexedEntriesIndex();
         }
-
-        entryUpdateDAO = handle.attach(EntryUpdateDAO.class);
-        entryUpdateDAO.ensureEntryTableInPlace();
-
-        itemUpdateDAO = handle.attach(ItemUpdateDAO.class);
-        itemUpdateDAO.ensureItemTableInPlace();
     }
 
     // TODO: Remove once migration to new schema complete
     public int lastReadSerialNumber() {
         return indexedEntriesUpdateDAO.lastReadSerialNumber();
-    }
-
-    public int lastReadEntryNumber() {
-        return entryUpdateDAO.lastReadEntryNumber();
     }
 
     @Transaction(TransactionIsolationLevel.SERIALIZABLE)
@@ -58,16 +45,6 @@ public abstract class DestinationDBUpdateDAO implements GetHandle, DBConnectionD
 
         upsertInCurrentKeysTable(registerName, orderedEntryIndex);
         indexedEntriesUpdateDAO.updateTotalEntries(orderedEntryIndex.size());
-    }
-
-    @Transaction(TransactionIsolationLevel.SERIALIZABLE)
-    public void writeEntriesAndItemsInBatch(List<Entry> entries, List<Item> items) {
-        Set<Item> newItems = extractNewItems(items);
-
-        entryUpdateDAO.writeBatch(entries);
-        if (!newItems.isEmpty()) {
-            itemUpdateDAO.writeBatch(newItems);
-        }
     }
 
     private void upsertInCurrentKeysTable(String registerName, List<OrderedEntryIndex> orderedEntryIndexes) {
@@ -93,8 +70,5 @@ public abstract class DestinationDBUpdateDAO implements GetHandle, DBConnectionD
         return jsonNode.get("entry").get(registerName).textValue();
     }
 
-    private Set<Item> extractNewItems(List<Item> items) {
-        List<String> existingItemHexValues = itemUpdateDAO.getExistingItemHexValues(Lists.transform(items, Item::getSha256hex));
-        return items.stream().filter(i -> !existingItemHexValues.contains(i.getSha256hex())).collect(Collectors.toSet());
-    }
+
 }
